pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mdlex"
version = "0.1.0"
authors = [
  { name="Matthias Luebken", email="first.last@gmail.com" },
]
description = "A utility for managing markdown documents with YAML frontmatter in SQLite"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "PyYAML>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/luebken/mdlex"
"Bug Tracker" = "https://github.com/luebken/mdlex/issues"

[project.scripts]
mdlex = "mdlex.__main__:main"